+++ src/main/java/org/folio/service/DeclareItemLostService.java
@@ -1,0 +1,7 @@
+package org.folio.service;
+
+import org.folio.domain.dto.DeclareItemLostRequest;
+
+public interface DeclareItemLostService {
+  void declareItemLost(DeclareItemLostRequest request);
+}